Add RandomUniform, move RandomNormal to fresh new ops directory
